world_mind-app/requirements.txt
streamlit
fastapi
uvicorn
langgraph
langchain
langchain-community
tavily-python
chromadb
sentence-transformers
python-dotenv
google-generativeai
